add basic game logic & game event handlers
add event handlers to display x or o on the game board, depending on the current player, and to alternate current player with each played turn.

add checkwinner function to run with each click handler to check if a winning combination has been made

when a winning combination has been made, successfully logged "winner!" to the console.